Refactored InventoryNumberComparer
Refactored the process, since the `InventoryNumber` is a string which usually contains integers we try to parse it to one, if this fails for either one of the two we compare them like strings.
This is not optimal I suggest the InventoryNumber to be changed to an integer instead of an id.